Cleanup for imports and file renaming
