+++ iris/api/core/screen.py
@@ -2,8 +2,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from util.image_search import SCREEN_WIDTH, SCREEN_HEIGHT
 from region import Region, get_region
+from util.image_search import SCREEN_WIDTH, SCREEN_HEIGHT
 
 
 # todo Refactor this
