+++ iris/api/core/settings.py
@@ -2,9 +2,10 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
+from platform import Platform
 from util.core_helper import get_os, get_os_version
 from util.parse_args import parse_args
-from platform_iris import Platform
+
 
 DEFAULT_MIN_SIMILARITY = 0.8
 DEFAULT_SLOW_MOTION_DELAY = 2
